chore(clerk-js): Remove BlockWithAction as it unused
