+++ packages/clerk-js/src/ui/components/OrganizationProfile/VerifyDomainPage.tsx
@@ -5,7 +5,6 @@ import { useCoreOrganization, useEnvironment } from '../../contexts';
 import { Button, descriptors, Flex, localizationKeys, Spinner } from '../../customizables';
 import type { VerificationCodeCardProps } from '../../elements';
 import {
-  BlockWithAction,
   ContentPage,
   Form,
   FormButtonContainer,
@@ -18,14 +17,13 @@ import { CodeForm } from '../../elements/CodeForm';
 import { useFetch, useLoadingStatus, useNavigateToFlowStart } from '../../hooks';
 import { useRouter } from '../../router';
 import { handleError, sleep, useFormControl } from '../../utils';
-import { EnrollmentBadge } from './EnrollmentBadge';
 import { OrganizationProfileBreadcrumbs } from './OrganizationProfileNavbar';
 
 export const VerifyDomainPage = withCardStateProvider(() => {
   const card = useCardState();
   const { organizationSettings } = useEnvironment();
   const { organization } = useCoreOrganization();
-  const { params, navigate } = useRouter();
+  const { params } = useRouter();
   const { navigateToFlowStart } = useNavigateToFlowStart();
 
   const [success, setSuccess] = React.useState(false);
@@ -134,23 +132,6 @@ export const VerifyDomainPage = withCardStateProvider(() => {
         headerSubtitle={subtitle}
         Breadcrumbs={OrganizationProfileBreadcrumbs}
       >
-        <BlockWithAction
-          elementDescriptor={descriptors.accordionTriggerButton}
-          badge={<EnrollmentBadge organizationDomain={domain} />}
-          sx={t => ({
-            backgroundColor: t.colors.$blackAlpha50,
-            padding: `${t.space.$3} ${t.space.$4}`,
-            minHeight: t.sizes.$10,
-          })}
-          actionLabel={localizationKeys('organizationProfile.verifyDomainPage.actionLabel__remove')}
-          onActionClick={() => navigate(`../../../domain/${domain.id}/remove`)}
-          actionSx={t => ({
-            color: t.colors.$danger500,
-          })}
-        >
-          {domain.name}
-        </BlockWithAction>
-
         <Form.Root onSubmit={onSubmitPrepare}>
           <Form.ControlRow elementId={emailField.id}>
             <Form.Control
